feat: migrate dog detail to screen header
