+++ apps/mobile/components/dogs/DogStatsCard.tsx
@@ -2,7 +2,7 @@ import { StyleSheet, Text, View } from 'react-native';
 import { useTranslation } from 'react-i18next';
 import { useColors } from '@/hooks/use-colors';
 import { spacing, typography } from '@/theme/tokens';
-import { OutlinedCard } from '@/components/ui/OutlinedCard';
+import { GroupedCard } from '@/components/ui/GroupedCard';
 import { formatDistance } from '@/lib/walk/format';
 import type { WalkStats } from '@/types/graphql';
 
@@ -16,17 +16,19 @@ export function DogStatsCard({ stats, streakDays = 0 }: DogStatsCardProps) {
   const theme = useColors();
 
   return (
-    <OutlinedCard style={styles.card}>
+    <GroupedCard elevated testID="dog-stats-card" style={styles.card}>
       <View style={styles.stat}>
         <Text style={[styles.value, { color: theme.onSurface }]}>{stats.totalWalks}</Text>
         <Text style={[styles.label, { color: theme.onSurfaceVariant }]}>{t('dogs.stats.walks')}</Text>
       </View>
+      <View testID="dog-stats-card-divider" style={[styles.divider, { backgroundColor: theme.border }]} />
       <View style={styles.stat}>
         <Text style={[styles.value, { color: theme.onSurface }]}>
           {formatDistance(stats.totalDistanceM, 'km', 1)}
         </Text>
         <Text style={[styles.label, { color: theme.onSurfaceVariant }]}>{t('dogs.stats.distance')}</Text>
       </View>
+      <View testID="dog-stats-card-divider" style={[styles.divider, { backgroundColor: theme.border }]} />
       <View style={styles.stat}>
         <Text style={[styles.value, { color: theme.onSurface }]}>
           {t('dogs.detail.streakDays', { days: streakDays })}
@@ -35,13 +37,15 @@ export function DogStatsCard({ stats, streakDays = 0 }: DogStatsCardProps) {
           {t('dogs.detail.streakLabel')}
         </Text>
       </View>
-    </OutlinedCard>
+    </GroupedCard>
   );
 }
 
 const styles = StyleSheet.create({
   card: {
     flexDirection: 'row',
+    paddingVertical: spacing.step14,
+    paddingHorizontal: spacing.xs,
     marginBottom: spacing.md,
   },
   stat: {
@@ -55,5 +59,9 @@ const styles = StyleSheet.create({
   label: {
     ...typography.metricLabel,
     marginTop: spacing.xs / 2,
+  },
+  divider: {
+    width: StyleSheet.hairlineWidth,
+    marginVertical: spacing.xs,
   },
 });
